backend_django/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.8.2
python-decouple==3.8
dj-database-url==2.1.0
